Move spec interface to api and add record docs
