--- asyncapi-core/src/main/java/io/ballerina/asyncapi/core/model/AsyncApiCorrelationId.java
@@ -1,10 +1,0 @@
-package io.ballerina.asyncapi.core.model;
-
-import java.util.Map;
-
-public record AsyncApiCorrelationId(
-        String description,
-        String location,
-        Map<String, String> extensions
-) {
-}
